# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * English Learning App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type RegisterRequestRole = typeof RegisterRequestRole[keyof typeof RegisterRequestRole];


export const RegisterRequestRole = {
  student: 'student',
  parent: 'parent',
} as const;

export interface RegisterRequest {
  username: string;
  password: string;
  name: string;
  role: RegisterRequestRole;
  /** Required for students (5-18) */
  age?: number;
  /** Parent user ID (optional, links student to parent) */
  parentId?: number;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  student: 'student',
  parent: 'parent',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  username: string;
  name: string;
  role: UserRole;
  age?: number | null;
  totalPoints: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserWithStats = User & ({
  totalTimeMinutes?: number;
  completedAssignments?: number;
  averageScore?: number | null;
});

export interface UpdateUserRequest {
  name?: string;
  age?: number;
}

export type AssignmentType = typeof AssignmentType[keyof typeof AssignmentType];


export const AssignmentType = {
  text_test: 'text_test',
  audio: 'audio',
  reading: 'reading',
  video: 'video',
} as const;

export interface Assignment {
  id: number;
  title: string;
  description: string;
  type: AssignmentType;
  ageMin: number;
  ageMax: number;
  points: number;
  mediaUrl?: string | null;
  createdAt: string;
}

export interface Question {
  id: number;
  text: string;
  options: string[];
  /** Only returned for admin role */
  correctAnswer?: string | null;
  orderIndex: number;
}

export type AssignmentDetail = Assignment & ({
  questions?: Question[];
  /** Reading text content for reading assignments */
  content?: string | null;
});

export type CreateAssignmentRequestType = typeof CreateAssignmentRequestType[keyof typeof CreateAssignmentRequestType];


export const CreateAssignmentRequestType = {
  text_test: 'text_test',
  audio: 'audio',
  reading: 'reading',
  video: 'video',
} as const;

export interface CreateQuestionRequest {
  text: string;
  options: string[];
  correctAnswer: string;
  orderIndex: number;
}

export interface CreateAssignmentRequest {
  title: string;
  description: string;
  type: CreateAssignmentRequestType;
  ageMin: number;
  ageMax: number;
  points: number;
  mediaUrl?: string;
  content?: string;
  questions?: CreateQuestionRequest[];
}

export interface UpdateAssignmentRequest {
  title?: string;
  description?: string;
  ageMin?: number;
  ageMax?: number;
  points?: number;
  mediaUrl?: string;
  content?: string;
}

export interface AnswerInput {
  questionId: number;
  answer: string;
}

export interface SubmitAssignmentRequest {
  answers: AnswerInput[];
  /** URL of student recording for reading assignments */
  recordingUrl?: string | null;
}

export interface QuestionResult {
  questionId: number;
  isCorrect: boolean;
  studentAnswer: string;
  correctAnswer: string;
}

export interface SubmissionResult {
  submissionId: number;
  score: number;
  totalQuestions: number;
  correctCount: number;
  pointsEarned: number;
  results: QuestionResult[];
}

export interface Submission {
  id: number;
  studentId: number;
  studentName: string;
  assignmentId: number;
  assignmentTitle: string;
  score: number;
  correctCount: number;
  totalQuestions: number;
  pointsEarned: number;
  recordingUrl?: string | null;
  submittedAt: string;
  answers: QuestionResult[];
}

export interface StudentError {
  assignmentId: number;
  assignmentTitle: string;
  questionText: string;
  studentAnswer: string;
  correctAnswer: string;
  occurredAt: string;
}

export interface VoiceChatSession {
  id: number;
  studentId: number;
  studentName: string;
  messageCount: number;
  pointsEarned: number;
  createdAt: string;
  updatedAt: string;
}

export type VoiceChatMessageRole = typeof VoiceChatMessageRole[keyof typeof VoiceChatMessageRole];


export const VoiceChatMessageRole = {
  student: 'student',
  ai: 'ai',
} as const;

export interface VoiceChatMessage {
  id: number;
  role: VoiceChatMessageRole;
  audioUrl?: string | null;
  transcript: string;
  createdAt: string;
}

export type VoiceChatSessionDetail = VoiceChatSession & {
  messages?: VoiceChatMessage[];
};

export interface SendVoiceMessageRequest {
  /** Base64 encoded audio data */
  audioBase64: string;
  /** Audio MIME type (e.g. audio/m4a, audio/webm) */
  mimeType: string;
}

export interface VoiceMessageResponse {
  studentMessage: VoiceChatMessage;
  aiMessage: VoiceChatMessage;
  pointsEarned: number;
}

export interface TimeSession {
  id: number;
  studentId: number;
  startedAt: string;
  endedAt?: string | null;
  durationMinutes?: number | null;
}

export interface TimeStats {
  totalMinutes: number;
  todayMinutes: number;
  weekMinutes: number;
  sessions: TimeSession[];
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  name: string;
  totalPoints: number;
  completedAssignments: number;
}

export interface UploadResult {
  url: string;
  filename: string;
}

export type ListUsersParams = {
role?: ListUsersRole;
parentId?: number;
};

export type ListUsersRole = typeof ListUsersRole[keyof typeof ListUsersRole];


export const ListUsersRole = {
  student: 'student',
  parent: 'parent',
  admin: 'admin',
} as const;

export type ListAssignmentsParams = {
type?: ListAssignmentsType;
ageMin?: number;
ageMax?: number;
};

export type ListAssignmentsType = typeof ListAssignmentsType[keyof typeof ListAssignmentsType];


export const ListAssignmentsType = {
  text_test: 'text_test',
  audio: 'audio',
  reading: 'reading',
  video: 'video',
} as const;

export type ListVoiceChatSessionsParams = {
studentId?: number;
};

export type UploadAudioBody = {
  file?: Blob;
};

export type UploadVideoBody = {
  file?: Blob;
};

export type UploadStudentRecordingBody = {
  file?: Blob;
  assignmentId?: number;
};